[level 0] Title: 다음에 올 숫자, Time: 0.00 ms, Memory: 10.3 MB -BaekjoonHub

+++ 프로그래머스/0/120924. 다음에 올 숫자/다음에 올 숫자.py
@@ -3,7 +3,7 @@ def solution(common):
         return common[-1] + (common[1] - common[0])
     else:
         ratio = common[1] / common[0]
-        return int(common[-1] * ratio)
+        return common[-1] * ratio
